calculo de huellaEcologica

+++ app/src/main/java/com/domain/kevin/myreportrash/ReporTrash_clases/HuellaEcologica.java
@@ -7,47 +7,47 @@ import android.os.Parcelable;
  * Created by kevin on 2/8/16.
  */
 public class HuellaEcologica implements Parcelable{
-    protected int comida;
-    protected int movilidad;
-    protected int agua;
-    protected int electricidad;
+    protected Double comida;
+    protected Double movilidad;
+    protected Double agua;
+    protected Double electricidad;
 
     public HuellaEcologica(){
 
     }
 
-    public HuellaEcologica(int comida, int movilidad, int agua, int electricidad){
+    public HuellaEcologica(Double comida, Double movilidad, Double agua, Double electricidad){
         this.comida = comida;
         this.movilidad = movilidad;
         this.agua = agua;
         this.electricidad = electricidad;
     }
 
-    public int getComida(){
+    public Double getComida(){
         return comida;
     }
-    public void setComida(int comida){
+    public void setComida(Double comida){
         this.comida = comida;
     }
 
-    public int getMovilidad(){
+    public Double getMovilidad(){
         return movilidad;
     }
-    public void setMovilidad(int movilidad){
+    public void setMovilidad(Double movilidad){
         this.movilidad = movilidad;
     }
 
-    public int getAgua(){
+    public Double getAgua(){
         return agua;
     }
-    public void setAgua(int agua){
+    public void setAgua(Double agua){
         this.agua = agua;
     }
 
-    public int getElectricidad(){
+    public Double getElectricidad(){
         return electricidad;
     }
-    public void setElectricidad(int electricidad){
+    public void setElectricidad(Double electricidad){
         this.electricidad = electricidad;
     }
 
@@ -58,18 +58,18 @@ public class HuellaEcologica implements Parcelable{
 
     @Override
     public void writeToParcel(Parcel dest, int flags) {
-        dest.writeInt(this.comida);
-        dest.writeInt(this.movilidad);
-        dest.writeInt(this.agua);
-        dest.writeInt(this.electricidad);
+        dest.writeDouble(this.comida);
+        dest.writeDouble(this.movilidad);
+        dest.writeDouble(this.agua);
+        dest.writeDouble(this.electricidad);
 
     }
 
     protected HuellaEcologica(Parcel in){
-        this.comida = in.readInt();
-        this.movilidad = in.readInt();
-        this.agua = in.readInt();
-        this.electricidad = in.readInt();
+        this.comida = in.readDouble();
+        this.movilidad = in.readDouble();
+        this.agua = in.readDouble();
+        this.electricidad = in.readDouble();
     }
     public static final Creator<HuellaEcologica> CREATOR = new Creator<HuellaEcologica>() {
         @Override
